Moving commands into their own folders, and changing the command list population method to make this easier
